DOP-4532 reconfigured how breadcrumbs are rendered
